agent-observability/backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# gRPC
grpcio==1.68.1
grpcio-tools==1.68.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20

# Config
pydantic-settings==2.7.0

# OpenTelemetry
opentelemetry-sdk==1.29.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0

# Utilities
httpx==0.28.1
